Convert PlotElement to frontend/backend
