+++ mfp/gui/clutter/__init__.py
@@ -1,7 +1,7 @@
 """
 Clutter UI backend
 
-Imports here make sure the backend is registered with the factory
+Imports here make sure the backends are registered
 """
 
 from .app_window import ClutterAppWindowBackend  # noqa
@@ -12,12 +12,13 @@ from .text_widget import ClutterTextWidgetImpl  # noqa
 from .base_element import ClutterBaseElementBackend  # noqa
 from .colordb import ClutterColorDBBackend  # noqa
 from .message_element import ClutterMessageElementImpl  # noqa
+from .plot_element import ClutterPlotElementImpl  # noqa
 from .processor_element import ClutterProcessorElementImpl  # noqa
 from .connection_element import ClutterConnectionElementImpl  # noqa
-from .slidemeter_element import (
-    ClutterFaderElementImpl,
-    ClutterBarMeterElementImpl,
-    ClutterDialElementImpl
+from .slidemeter_element import (  # noqa
+    ClutterFaderElementImpl,  # noqa
+    ClutterBarMeterElementImpl,  # noqa
+    ClutterDialElementImpl  # noqa
 )
 from .via_element import (  # noqa
     ClutterSendViaElementImpl,  # noqa
@@ -25,11 +26,11 @@ from .via_element import (  # noqa
     ClutterSendSignalViaElementImpl,  # noqa
     ClutterReceiveSignalViaElementImpl,  # noqa
 )
-from .text_element import ClutterTextElementImpl
-from .enum_element import ClutterEnumElementImpl
-from .button_element import (
-    ClutterButtonElementImpl,
-    ClutterToggleIndicatorElementImpl,
-    ClutterToggleButtonElementImpl,
-    ClutterBangButtonElementImpl,
+from .text_element import ClutterTextElementImpl  # noqa
+from .enum_element import ClutterEnumElementImpl  # noqa
+from .button_element import (  # noqa
+    ClutterButtonElementImpl,  # noqa
+    ClutterToggleIndicatorElementImpl,  # noqa
+    ClutterToggleButtonElementImpl,  # noqa
+    ClutterBangButtonElementImpl,  # noqa
 )
