+++ mfp/gui/clutter/xyplot/scatterplot.py
