+++ mfp/gui/clutter/xyplot/scopeplot.py
